+++ awc/examples/client.rs
@@ -1,18 +1,20 @@
-use actix_http::Error;
+use std::error::Error as StdError;
 
 #[actix_web::main]
-async fn main() -> Result<(), Error> {
-    std::env::set_var("RUST_LOG", "actix_http=trace");
+async fn main() -> Result<(), Box<dyn StdError>> {
+    std::env::set_var("RUST_LOG", "client=trace,awc=trace,actix_http=trace");
     env_logger::init();
 
     let client = awc::Client::new();
 
     // Create request builder, configure request and send
-    let mut response = client
+    let request = client
         .get("https://www.rust-lang.org/")
-        .append_header(("User-Agent", "Actix-web"))
-        .send()
-        .await?;
+        .append_header(("User-Agent", "Actix-web"));
+
+    println!("Request: {:?}", request);
+
+    let mut response = request.send().await?;
 
     // server http response
     println!("Response: {:?}", response);