+++ src/middleware/compress.rs
@@ -13,7 +13,6 @@ use actix_http::{
     body::{MessageBody, ResponseBody},
     encoding::Encoder,
     http::header::{ContentEncoding, ACCEPT_ENCODING},
-    Error,
 };
 use actix_service::{Service, Transform};
 use actix_utils::future::{ok, Ready};
@@ -23,6 +22,7 @@ use pin_project::pin_project;
 use crate::{
     dev::BodyEncoding,
     service::{ServiceRequest, ServiceResponse},
+    Error,
 };
 
 /// Middleware for compressing response payloads.